+++ bull-common/src/main/java/com/hotels/transformer/base/Defaults.java
@@ -27,15 +27,13 @@ import static com.hotels.transformer.utils.ClassUtils.isInt;
 import static com.hotels.transformer.utils.ClassUtils.isLong;
 import static com.hotels.transformer.utils.ClassUtils.isShort;
 
-import static lombok.AccessLevel.PRIVATE;
-
-import lombok.NoArgsConstructor;
+import lombok.experimental.UtilityClass;
 
 /**
  * Default values for primitive types.
  */
-@NoArgsConstructor(access = PRIVATE)
-public final class Defaults {
+@UtilityClass
+public class Defaults {
     /**
      * Gets the default value of a primitive type.
      * @param type the class type